+++ src/net.hpp
@@ -9,14 +9,14 @@ class Net
 {
 private:
     vector<Layer> m_layers; //m_layers[layerNum][neuronNum]
-    double m_error;
-    double m_recentAverageError;
-    static double m_recentAverageSmoothingFactor;
+    float m_error;
+    float m_recentAverageError;
+    static float m_recentAverageSmoothingFactor;
 
 public:
     Net(const vector<unsigned> &topology);
-    void feedForward(const vector<double> &inputVals);
-    void backProp(const vector<double> &targetVals);
-    void getResults(vector<double> &resultVals) const;
-    double getRecentAverageError(void) const {return m_recentAverageError;};
+    void feedForward(const vector<float> &inputVals);
+    void backProp(const vector<float> &targetVals);
+    void getResults(vector<float> &resultVals) const;
+    float getRecentAverageError(void) const {return m_recentAverageError;};
 };
